refactor: hoist collections data
