+++ packages/mask/src/extension/popups/pages/Wallet/components/WalletAssets/index.tsx
@@ -1,8 +1,8 @@
 import { Icons } from '@masknet/icons'
-import { RestorableScroll } from '@masknet/shared'
+import { CollectionList, RestorableScroll, UserAssetsProvider } from '@masknet/shared'
 import { NetworkPluginID, PopupRoutes } from '@masknet/shared-base'
 import { makeStyles, useTabs } from '@masknet/theme'
-import { useWallet } from '@masknet/web3-hooks-base'
+import { useAccount, useWallet } from '@masknet/web3-hooks-base'
 import { TabContext, TabList, TabPanel } from '@mui/lab'
 import { Box, Button, Tab, styled, tabClasses, tabsClasses } from '@mui/material'
 import { memo, useCallback } from 'react'
@@ -14,7 +14,13 @@ import { WalletContext } from '../../hooks/useWalletContext.js'
 import { ActivityList } from '../ActivityList/index.js'
 import { AssetsList } from '../AssetsList/index.js'
 import { WalletAssetTabs } from '../../type.js'
-import { Collectibles } from '../Collectibles/index.js'
+import type { Web3Helper } from '@masknet/web3-helpers'
+import urlcat from 'urlcat'
+
+const gridProps = {
+    columns: 'repeat(auto-fill, minmax(20%, 1fr))',
+    gap: '8px',
+}
 
 const useStyles = makeStyles()((theme) => {
     const isDark = theme.palette.mode === 'dark'
@@ -110,6 +116,7 @@ export interface WalletAssetsUIProps {
 
 export const WalletAssetsUI = memo<WalletAssetsUIProps>(function WalletAssetsUI({ onAddToken }) {
     const { t } = useI18N()
+    const navigate = useNavigate()
     const [params, setParams] = useSearchParams()
     const paramTab = params.get('tab') as WalletAssetTabs
 
@@ -121,6 +128,30 @@ export const WalletAssetsUI = memo<WalletAssetsUIProps>(function WalletAssetsUI(
         WalletAssetTabs.Activity,
     )
 
+    const account = useAccount(NetworkPluginID.PLUGIN_EVM)
+    const SEARCH_KEY = 'collectionId'
+
+    const handleItemClick = useCallback(
+        (asset: Web3Helper.NonFungibleTokenAll) => {
+            const path = urlcat(PopupRoutes.CollectibleDetail, {
+                chainId: asset.chainId,
+                address: asset.address,
+                id: asset.tokenId,
+            })
+            navigate(path, { state: { asset } })
+        },
+        [navigate],
+    )
+    const handleCollectionChange = useCallback((id: string | undefined) => {
+        setParams(
+            (params) => {
+                if (!id) params.delete(SEARCH_KEY)
+                else params.set(SEARCH_KEY, id)
+                return params.toString()
+            },
+            { replace: true },
+        )
+    }, [])
     return (
         <div className={classes.content}>
             <TabContext value={currentTab}>
@@ -157,23 +188,36 @@ export const WalletAssetsUI = memo<WalletAssetsUIProps>(function WalletAssetsUI(
                         <Icons.AddNoBorder size={16} />
                     </Button>
                 </Box>
-                <Box className={classes.panels}>
-                    <RestorableScroll scrollKey="assets" enabled={currentTab === WalletAssetTabs.Tokens}>
-                        <TabPanel value={WalletAssetTabs.Tokens} className={classes.tabPanel}>
-                            <AssetsList />
-                        </TabPanel>
-                    </RestorableScroll>
-                    <RestorableScroll scrollKey="collectibles" enabled={currentTab === WalletAssetTabs.Collectibles}>
-                        <TabPanel value={WalletAssetTabs.Collectibles} className={classes.tabPanel}>
-                            <Collectibles />
-                        </TabPanel>
-                    </RestorableScroll>
-                    <RestorableScroll scrollKey="activities" enabled={currentTab === WalletAssetTabs.Activity}>
-                        <TabPanel value={WalletAssetTabs.Activity} className={classes.tabPanel}>
-                            <ActivityList />
-                        </TabPanel>
-                    </RestorableScroll>
-                </Box>
+
+                <UserAssetsProvider
+                    pluginID={NetworkPluginID.PLUGIN_EVM}
+                    account={account}
+                    defaultCollectionId={params.get(SEARCH_KEY) || undefined}>
+                    <Box className={classes.panels}>
+                        <RestorableScroll scrollKey="assets" enabled={currentTab === WalletAssetTabs.Tokens}>
+                            <TabPanel value={WalletAssetTabs.Tokens} className={classes.tabPanel}>
+                                <AssetsList />
+                            </TabPanel>
+                        </RestorableScroll>
+                        <RestorableScroll
+                            scrollKey="collectibles"
+                            enabled={currentTab === WalletAssetTabs.Collectibles}>
+                            <TabPanel value={WalletAssetTabs.Collectibles} className={classes.tabPanel}>
+                                <CollectionList
+                                    gridProps={gridProps}
+                                    disableSidebar
+                                    onItemClick={handleItemClick}
+                                    onCollectionChange={handleCollectionChange}
+                                />
+                            </TabPanel>
+                        </RestorableScroll>
+                        <RestorableScroll scrollKey="activities" enabled={currentTab === WalletAssetTabs.Activity}>
+                            <TabPanel value={WalletAssetTabs.Activity} className={classes.tabPanel}>
+                                <ActivityList />
+                            </TabPanel>
+                        </RestorableScroll>
+                    </Box>
+                </UserAssetsProvider>
             </TabContext>
         </div>
     )
